Update README.md
In this Python-Flask-Blog Project, we are going to create posts on blog and we can manage how many posts
will going to display on the home page. Admin can add new posts and can manage all the posts and
performs different operations on it. I created a session variable because we don’t want our user to login
again and again and if user in session is true, then only user returns to Dashboard page.